Create `evmil::asm` module
This groups together various related aspects of the library.
