+++ src/asm/instruction.rs
@@ -9,7 +9,7 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-use super::{AbstractInstruction,Instruction,InstructionOperands};
+use crate::instruction::{AbstractInstruction,Instruction,InstructionOperands};
 
 // ============================================================================
 // Concrete Instructions
