--- src/assembler/mod.rs
@@ -1,9 +1,0 @@
-// Responsible for code generation.
-mod codegen;
-// Support for parsing.
-mod lexer;
-// Responsible for parsing.
-mod parser;
-
-pub use parser::{parse,AssemblyError};
-pub use codegen::{assemble,AssembleError};
