+++ src/bin/evmil.rs
@@ -18,8 +18,8 @@ use log4rs::append::console::ConsoleAppender;
 use log4rs::config::{Appender, Config, Root};
 use log4rs::encode::pattern::PatternEncoder;
 //
-use evmil::bytecode::{Assembly,Section};
-use evmil::instruction::{AssemblyInstruction};
+use evmil::asm::{Assembly,AssemblyInstruction};
+use evmil::bytecode::{Section};
 use evmil::{eof,legacy};
 use evmil::il::{Compiler,Parser};
 use evmil::util::{FromHexString, ToHexString};
