+++ src/legacy.rs
@@ -11,8 +11,9 @@
 // limitations under the License.
 use std::collections::HashMap;
 use crate::util::{Concretizable,w256,IsBottom,Top};
+use crate::asm::{AssemblyInstruction};
 use crate::bytecode::{Bytecode,Section};
-use crate::instruction::{AssemblyInstruction,Instruction,ToInstructions};
+use crate::instruction::{Instruction,ToInstructions};
 use crate::instruction::AbstractInstruction::*;
 use crate::execution::{Execution,ExecutionSection};
 use crate::state::{EvmState,EvmMemory,EvmStack,EvmStorage,EvmWord};
